changed ALOT of security related shit

+++ controllers/gl/bank_account_reconcile.php
@@ -17,6 +17,9 @@
       $this->Dates = Dates::i();
       $this->JS->_openWindow(800, 500);
       $this->JS->_footerFile('/js/reconcile.js');
+      if (!Input::post('bank_account')) {
+        $_POST['bank_account'] = 5;
+      }
       if ($this->Input->_post('reset')) {
         // GL_Account::reset_sql_for_reconcile($_POST['bank_account'], $this->Input->_post('reconcile_date'));
         $this->updateData();
@@ -107,59 +110,55 @@
       if ($_POST['bank_date']) {
         $sql .= " OR bt.reconciled='" . $_POST['bank_date'] . "'";
       }
-      $sql .= ") AND bt.amount!=0 GROUP BY bt.id ORDER BY bt.reconciled DESC,trans_date  DESC, amount ";
-
+      $sql .= ") AND bt.amount!=0 GROUP BY bt.id ORDER BY IF(bt.trans_date>='$begin' AND bt.trans_date<='$end',1,0) , bt.reconciled DESC    ,bt.trans_date  , amount ";
       $this->DB->_query($sql);
-      $rec    = $this->DB->_fetchAll();
-      $sql    = "SELECT date as state_date, amount as state_amount,memo FROM temprec WHERE  date >= '$begin' AND  date <='" . $end . "' ORDER BY date DESC,amount";
+      $rec = $this->DB->_fetchAll();
+      $sql = "SELECT date as state_date, amount as state_amount,memo FROM temprec WHERE  date >= '$begin' AND  date <='" . $end . "' ORDER BY date  ,amount";
       $this->DB->_query($sql);
-      $state      = $this->DB->_fetchAll();
-      if (!$state) return $this->oldWay();
-      $recced     = $unrecced = [];
-      $emptyrec   = [
+      $statement_trans = $this->DB->_fetchAll();
+      if (!$statement_trans) {
+        return $this->oldWay();
+      }
+      $known_trans                 = [];
+      $known_headers               = [
         'type', 'trans_no', 'ref', 'trans_date', 'id', 'amount', 'person_id', 'person_type_id', 'reconciled'
       ];
-      $emptyrec   = array_combine(array_values($emptyrec), array_pad([], count($emptyrec), ''));
-      $emptystate = array_combine(array_keys($state[0]), array_values(array_pad([], count($state[0]), '')));
-      while ($v = array_pop($state)) {
+      $known_headers               = array_combine(array_values($known_headers), array_pad([], count($known_headers), ''));
+      $statement_transment_headers = array_combine(array_keys($statement_trans[0]), array_values(array_pad([], count($statement_trans[0]), '')));
+      while ($v = array_shift($statement_trans)) {
         $amount = $v['state_amount'];
         foreach ($rec as $p=> $q) {
           if ($q['amount'] == $amount) {
             $matched = $rec[$p] + $v;
             unset($rec[$p]);
-            $recced[] = $matched;
+            $known_trans[] = $matched;
             continue 2;
           }
         }
-        $newv = $emptyrec;
+        $newv = $known_headers;
         Arr::append($newv, $v);
-        $recced[] = $newv;
+        $known_trans[] = $newv;
       }
       foreach ($rec as &$r) {
-        Arr::append($r, $emptystate);
+        Arr::append($r, $statement_transment_headers);
       }
-      Arr::append($recced, $rec);
-      usort($recced, [$this, 'sortByOrder']);
-      $cols = [
-        'Type'              => [
-          'fun'=> array($this, 'sysTypeName')
-        ], '#'              => [
-          'align'=> 'center', 'fun'=> array($this, 'viewTrans')
-        ], 'Ref'            => [
-          'fun'=> function($row) {
-            return substr($row['ref'], 0, 6);
-          }
-        ], 'Date'           => ['type'=> 'date'], 'Debit'       => [
-          'align'=> 'right', 'fun'=> array($this, 'formatDebit')
-        ], 'Credit'         => [
-          'align'=> 'right', 'insert'=> true, 'fun'=> array($this, 'formatCredit')
-        ], 'Info'           => ['fun'=> array($this, 'formatPerson')], 'GL'          => [
-          'fun'=> array($this, 'viewGl')
-        ], ''               => [
-          'fun'=> array($this, 'reconcileCheckbox')
-        ], 'Bank Date'      => ['type'=> 'date'], 'Amount'=> ['align'=> 'right', 'class'=> 'bold'], 'Info'
+      Arr::append($known_trans, $rec);
+      usort($known_trans, [$this, 'sortByOrder']);
+      $cols            = [
+        'Type'              => ['fun'=> array($this, 'sysTypeName')], //
+        '#'                 => ['align'=> 'center', 'fun'=> array($this, 'viewTrans')], //
+        'Ref'               => ['fun'=> 'formatReference'], //
+        'Date'              => ['type'=> 'date'], //
+        'Debit'             => ['align'=> 'right', 'fun'=> array($this, 'formatDebit')], //
+        'Credit'            => ['align'=> 'right', 'insert'=> true, 'fun'=> array($this, 'formatCredit')], //
+        'Info'              => ['fun'=> array($this, 'formatInfo')], //
+        'GL'                => ['fun'=> array($this, 'viewGl')], //
+        ['fun'=> array($this, 'reconcileCheckbox')], //
+        'Bank Date'         => ['type'=> 'date'], //
+        'Amount'            => ['align'=> 'right', 'class'=> 'bold'], //
+        'Info'
       ];
-      $table           = DB_Pager::new_db_pager('bank_rec', $recced, $cols);
+      $table           = DB_Pager::new_db_pager('bank_rec', $known_trans, $cols);
       $table->rowClass = function($row) {
         if (($row['trans_date'] && $row['reconciled'] && !$row['state_date']) || ($row['state_date'] && !$row['reconciled'])) {
           return "overduebg";
@@ -178,7 +177,7 @@
       Table::start();
       Table::header(_("Reconcile Date"));
       Row::start();
-      $_POST['reconcile_date']=$this->Dates->_sqlToDate($_POST['reconcile_date']);
+      $_POST['reconcile_date'] = $this->Dates->_sqlToDate($_POST['reconcile_date']);
       Forms::dateCells("", "reconcile_date", _('Date of bank statement to reconcile'), $this->Input->_post('bank_date') == '', 0, 0, 0, null, true);
       Row::end();
       Table::header(_("Beginning Balance"));
@@ -210,7 +209,24 @@
      * @return int
      */
     protected function getTotal() {
-      $total  = 0;
+      $total = 0;
+      if ($_POST['bank_account'] == 5) {
+        $date = $_POST['bank_date'];
+        if ($date) {
+          $begin = $this->Dates->_dateToSql($this->Dates->_beginMonth($date));
+          $end   = $this->Dates->_dateToSql($this->Dates->_endMonth($date));
+        } else {
+          $begin = "(SELECT max(reconciled) from bank_trans)";
+          $end   = $this->Dates->_today();
+        }
+        $sql                  = "(select (rb - amount) as amount from temprec where date>='$begin' and date<='$end' order by id desc limit 0,1) union (select rb as amount from temprec where date>='$begin' and date<='$end' order by id asc limit 0,1)";
+        $result               = $this->DB->_query($sql);
+        $beg_balance = $this->DB->_fetch($result)['amount'];
+          $end_balance = $this->DB->_fetch($result)['amount'];
+        $_POST["beg_balance"] = Num::priceFormat($beg_balance);
+        $_POST["end_balance"] = Num::priceFormat($end_balance);
+        $_POST["reconciled"]  = Num::priceFormat($end_balance- $beg_balance);
+      }
       $result = GL_Account::get_max_reconciled($this->Input->_post('reconcile_date'), $_POST['bank_account']);
       if ($row = $this->DB->_fetch($result)) {
         $_POST["reconciled"] = Num::priceFormat($row["end_balance"] - $row["beg_balance"]);
@@ -233,7 +249,7 @@
     /**
      * @return bool
      */
-    function check_date() {
+    function checkDate() {
       if (!$this->Dates->_isDate($this->Input->_post('reconcile_date'))) {
         Event::error(_("Invalid reconcile date format"));
         $this->JS->_setFocus('reconcile_date');
@@ -296,8 +312,10 @@
         return $this->ungroupButton($row);
       }
       $content = GL_UI::viewTrans($row["type"], $row["trans_no"]);
-      $content .= '<br><a href="' . e('/system/void_transaction?type=' . $row['type'] . '&trans_no=' . $row['trans_no'] . '&memo=Deleted during reconcile.') . '" target="_blank"
+      if (!$row['reconciled']) {
+        $content .= '<br><a href="' . e('/system/void_transaction?type=' . $row['type'] . '&trans_no=' . $row['trans_no'] . '&memo=Deleted during reconcile.') . '" target="_blank"
                                     class="button">void</a>';
+      }
       return $content;
     }
     /**
@@ -340,7 +358,7 @@
      *
      * @return string
      */
-    function formatPerson($row) {
+    function formatInfo($row) {
       if ($row['type'] == ST_BANKTRANSFER) {
         return DB_Comments::get_string(ST_BANKTRANSFER, $row['trans_no']);
       } elseif ($row['type'] == ST_GROUPDEPOSIT) {
@@ -370,7 +388,7 @@
      * @return bool
      */
     function change_tpl_flag($reconcile_id) {
-      if (!$this->check_date() && Forms::hasPost("rec_" . $reconcile_id)) // temporary fix
+      if (!$this->checkDate() && Forms::hasPost("rec_" . $reconcile_id)) // temporary fix
       {
         return false;
       }
@@ -381,7 +399,7 @@
       $_POST['bank_date'] = $this->Dates->_dateToSql($this->Input->_post('reconcile_date'));
       $reconcile_value    = Forms::hasPost("rec_" . $reconcile_id) ? ("'" . $_POST['bank_date'] . "'") : 'null';
       GL_Account::update_reconciled_values($reconcile_id, $reconcile_value, $_POST['reconcile_date'], Validation::input_num('end_balance'), $_POST['bank_account']);
-      $this->Ajax->_activate('summary');
+      $this->Ajax->_activate('_page_body');
       $this->JS->_setFocus($reconcile_id);
       return true;
     }
@@ -403,13 +421,21 @@
      */
     public function sortByOrder($a, $b) {
       $date1 = $a['state_date'] ? : $a['trans_date'];
-            $date2 = $b['state_date'] ? : $b['trans_date'];
-      if ($date1==$date2) {
+      $date2 = $b['state_date'] ? : $b['trans_date'];
+      if ($date1 == $date2) {
         $amount1 = $a['state_amount'] ? : $a['amount'];
-              $amount2 = $b['state_amount'] ? : $b['amount'];
-        return $amount1-$amount2;
+        $amount2 = $b['state_amount'] ? : $b['amount'];
+        return $amount1 - $amount2;
       }
-      return strcmp($date1,$date2);
+      return strcmp($date1, $date2);
+    }
+    /**
+     * @param $row
+     *
+     * @return string
+     */
+    public function formatReference($row) {
+      return substr($row['ref'], 0, 7);
     }
     /**
      * @return bool
@@ -422,15 +448,11 @@
       $cols         = array(
         _("Type")        => array('fun' => array($this, 'sysTypeName'), 'ord' => ''), //
         _("#")           => array('fun' => array($this, 'viewTrans'), 'ord' => ''), //
-        _("Reference")   => array(
-          'fun'=> function($row) {
-            return substr($row['ref'], 0, 6);
-          }
-        ), //
+        _("Reference")   => array('fun'=> [$this, 'formatReference']), //
         _("Date")        => array('type'=> 'date', 'ord' => ''), //
         _("Debit")       => array('align' => 'right', 'fun' => array($this, 'formatDebit'), 'ord' => ''), //
         _("Credit")      => array('align' => 'right', 'insert' => true, 'fun' => array($this, 'formatCredit'), 'ord' => ''), //
-        _("Person/Item") => array('fun' => array($this, 'formatPerson')), //
+        _("Person/Item") => array('fun' => array($this, 'formatInfo')), //
         array('insert' => true, 'fun' => array($this, 'viewGl')), //
         "X"              => array('insert' => true, 'fun' => array($this, 'reconcileCheckbox')), //
         array('insert' => true, 'fun' => array($this, 'ungroupButton'))
